Refactor Dashboard into digestible pieces
