+++ src/components/Dashboard.tsx
@@ -1,318 +1,14 @@
 import { useState, useMemo } from 'react';
-import {
-  ExternalLink,
-  Star,
-  GitFork,
-  Users,
-  Clock,
-  ChevronDown,
-  ChevronUp,
-  ChevronLeft,
-  ChevronRight,
-  Copy,
-  Check,
-  Search,
-  Filter,
-} from 'lucide-react';
-import { PlatformLogo } from './PlatformLogo';
-import type { Repository, OwnerInfo, Metrics } from '../types';
 import repoData from '../data/repo-data.json';
+import { RepoCard } from './RepoCard';
+import { Pagination } from './Pagination';
+import { SearchAndFilter } from './SearchAndFilter';
+import { useRepositoryFiltering } from '../hooks/useRepositoryFiltering';
+import type { Repository } from '../types';
 
-const OwnerCard = ({ owner }: { owner: OwnerInfo }) => {
-  const [isExpanded, setIsExpanded] = useState(false);
-  const [copied, setCopied] = useState(false);
-
-  const handleCopyEmail = async () => {
-    if (owner.email) {
-      await navigator.clipboard.writeText(owner.email);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 1500);
-    }
-  };
-
-  return (
-    <div className="mt-4 bg-gray-50 rounded-lg p-4">
-      <div
-        className="flex items-center justify-between cursor-pointer"
-        onClick={() => setIsExpanded(!isExpanded)}
-      >
-        <div className="flex items-center gap-2">
-          <div className="text-sm font-medium">{owner.display_name}</div>
-          <span className="text-xs px-2 py-1 bg-gray-200 rounded-full">{owner.type}</span>
-        </div>
-        {isExpanded ? <ChevronUp size={16} /> : <ChevronDown size={16} />}
-      </div>
-
-      {isExpanded && (
-        <div className="mt-4 space-y-2 text-sm">
-          {owner.bio && <p className="text-gray-600">{owner.bio}</p>}
-          <div className="grid grid-cols-2 gap-4">
-            {owner.location && (
-              <div>
-                <span className="font-medium">Location:</span> {owner.location}
-              </div>
-            )}
-            {owner.company && (
-              <div>
-                <span className="font-medium">Company:</span> {owner.company}
-              </div>
-            )}
-            {owner.public_repos !== undefined && (
-              <div>
-                <span className="font-medium">Public Repos:</span> {owner.public_repos}
-              </div>
-            )}
-            {owner.followers !== undefined && (
-              <div>
-                <span className="font-medium">Followers:</span> {owner.followers}
-              </div>
-            )}
-            {owner.members_count !== undefined && (
-              <div>
-                <span className="font-medium">Members:</span> {owner.members_count}
-              </div>
-            )}
-          </div>
-          {(owner.social_links.website || owner.social_links.twitter || owner.email) && (
-            <div className="flex gap-4 mt-2">
-              {owner.email && (
-                <div className="flex items-center gap-2">
-                  <a
-                    href={`mailto:${owner.email}`}
-                    rel="noopener noreferrer"
-                    className="text-blue-600 hover:text-blue-800 flex items-center gap-1"
-                  >
-                    <ExternalLink size={14} />
-                    Email
-                  </a>
-                  <button
-                    onClick={handleCopyEmail}
-                    className="p-1 hover:bg-gray-200 rounded-full transition-colors"
-                    title={copied ? 'Copied!' : `Copy email: ${owner.email}`}
-                  >
-                    {copied ? <Check size={14} className="text-green-500" /> : <Copy size={14} />}
-                  </button>
-                </div>
-              )}
-              {owner.social_links.website && (
-                <a
-                  href={owner.social_links.website}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="text-blue-600 hover:text-blue-800 flex items-center gap-1"
-                >
-                  <ExternalLink size={14} />
-                  Website
-                </a>
-              )}
-              {owner.social_links.twitter && (
-                <a
-                  href={owner.social_links.twitter}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="text-blue-600 hover:text-blue-800 flex items-center gap-1"
-                >
-                  <ExternalLink size={14} />
-                  Twitter
-                </a>
-              )}
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
-};
-
-const MetricsBar = ({ metrics }: { metrics: Metrics }) => {
-  return (
-    <div className="flex gap-4 mt-4 text-sm text-gray-600">
-      <div className="flex items-center gap-1">
-        <Star size={16} className="text-yellow-500" />
-        <span>{metrics.stars}</span>
-      </div>
-      <div className="flex items-center gap-1">
-        <GitFork size={16} className="text-gray-500" />
-        <span>{metrics.forks}</span>
-      </div>
-      <div className="flex items-center gap-1">
-        <Users size={16} className="text-gray-500" />
-        <span>{metrics.contributors.total_count}</span>
-      </div>
-    </div>
-  );
-};
-
-const RepoCard = ({ repo }: { repo: Repository }) => {
-  const formatDate = (dateString: string) => {
-    return new Date(dateString).toLocaleDateString('en-US', {
-      year: 'numeric',
-      month: 'short',
-      day: 'numeric',
-    });
-  };
-
-  const getPlatformColor = (platform: string) => {
-    switch (platform.toLowerCase()) {
-      case 'github':
-        return 'text-gray-900';
-      case 'gitlab':
-        return 'text-orange-600';
-      default:
-        return 'text-gray-600';
-    }
-  };
-
-  return (
-    <div className="bg-white rounded-lg shadow-md p-6 hover:shadow-lg transition-shadow">
-      <div className="flex items-start justify-between">
-        <div>
-          <a
-            href={repo.metadata.url}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="text-lg font-semibold text-blue-600 hover:text-blue-800 flex items-center gap-2"
-          >
-            {repo.metadata.name}
-            <ExternalLink size={16} />
-          </a>
-          <p className="text-sm text-gray-600 mt-2">{repo.metadata.description}</p>
-          {repo.metadata.categories && (
-            <div className="flex flex-wrap gap-2 mt-2">
-              {repo.metadata.categories.map((category) => (
-                <span
-                  key={category}
-                  className="text-xs px-2 py-1 bg-blue-100 text-blue-800 rounded-full"
-                >
-                  {category}
-                </span>
-              ))}
-            </div>
-          )}
-        </div>
-        <span
-          className={`flex items-center gap-1 text-xs px-2 py-1 bg-gray-100 rounded-full ${getPlatformColor(repo.platform)}`}
-        >
-          <PlatformLogo platform={repo.platform} className="w-4 h-4" />
-          {repo.platform}
-        </span>
-      </div>
-
-      <MetricsBar metrics={repo.metrics} />
-
-      <div className="flex items-center gap-2 mt-4 text-sm text-gray-500">
-        <Clock size={16} />
-        <span>Updated {formatDate(repo.metadata.last_updated)}</span>
-      </div>
-
-      <OwnerCard owner={repo.owner_info} />
-    </div>
-  );
-};
-
-const Pagination = ({
-  currentPage,
-  totalPages,
-  onPageChange,
-}: {
-  currentPage: number;
-  totalPages: number;
-  onPageChange: (page: number) => void;
-}) => {
-  return (
-    <div className="flex items-center justify-center gap-2 mt-8">
-      <button
-        onClick={() => onPageChange(currentPage - 1)}
-        disabled={currentPage === 1}
-        className="p-2 rounded-lg hover:bg-gray-200 disabled:opacity-50 disabled:hover:bg-transparent"
-      >
-        <ChevronLeft size={20} />
-      </button>
-
-      <div className="flex items-center gap-1">
-        {Array.from({ length: totalPages }, (_, i) => i + 1).map((page) => (
-          <button
-            key={page}
-            onClick={() => onPageChange(page)}
-            className={`px-3 py-1 rounded-lg ${
-              currentPage === page ? 'bg-blue-600 text-white' : 'hover:bg-gray-200'
-            }`}
-          >
-            {page}
-          </button>
-        ))}
-      </div>
-
-      <button
-        onClick={() => onPageChange(currentPage + 1)}
-        disabled={currentPage === totalPages}
-        className="p-2 rounded-lg hover:bg-gray-200 disabled:opacity-50 disabled:hover:bg-transparent"
-      >
-        <ChevronRight size={20} />
-      </button>
-
-      <div className="ml-4 text-sm text-gray-600">
-        Page {currentPage} of {totalPages}
-      </div>
-    </div>
-  );
-};
-
-const SearchAndFilter = ({
-  searchTerm,
-  onSearchChange,
-  selectedCategories,
-  onCategoryChange,
-  categories,
-}: {
-  searchTerm: string;
-  onSearchChange: (term: string) => void;
-  selectedCategories: string[];
-  onCategoryChange: (category: string) => void;
-  categories: string[];
-}) => {
-  return (
-    <div className="mb-8 space-y-4">
-      <div className="relative">
-        <Search
-          className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400"
-          size={20}
-        />
-        <input
-          type="text"
-          placeholder="Search projects..."
-          value={searchTerm}
-          onChange={(e) => onSearchChange(e.target.value)}
-          className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
-        />
-      </div>
-      <div>
-        <div className="flex items-center gap-2 mb-2">
-          <Filter size={20} className="text-gray-600" />
-          <span className="font-medium text-gray-700">Categories</span>
-        </div>
-        <div className="flex flex-wrap gap-2">
-          {categories.map((category) => (
-            <button
-              key={category}
-              onClick={() => onCategoryChange(category)}
-              className={`px-3 py-1 rounded-full text-sm ${
-                selectedCategories.includes(category)
-                  ? 'bg-blue-600 text-white'
-                  : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
-              }`}
-            >
-              {category}
-            </button>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-};
+const ITEMS_PER_PAGE = 10;
 
 const Dashboard = () => {
-  const ITEMS_PER_PAGE = 10;
   const [currentPage, setCurrentPage] = useState(1);
   const [searchTerm, setSearchTerm] = useState('');
   const [selectedCategories, setSelectedCategories] = useState<string[]>([]);
@@ -327,20 +23,11 @@ const Dashboard = () => {
   }, []);
 
   // Filter repositories based on search term and categories
-  const filteredRepos = useMemo(() => {
-    return repoData.repositories.filter((repo) => {
-      const matchesSearch =
-        searchTerm === '' ||
-        repo.metadata.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-        repo.metadata.description?.toLowerCase().includes(searchTerm.toLowerCase());
-
-      const matchesCategories =
-        selectedCategories.length === 0 ||
-        selectedCategories.every((category) => repo.metadata.categories?.includes(category));
-
-      return matchesSearch && matchesCategories;
-    });
-  }, [searchTerm, selectedCategories]);
+  const filteredRepos = useRepositoryFiltering(
+    repoData.repositories,
+    searchTerm,
+    selectedCategories
+  );
 
   const totalPages = Math.ceil(filteredRepos.length / ITEMS_PER_PAGE);
   const startIndex = (currentPage - 1) * ITEMS_PER_PAGE;
@@ -388,7 +75,7 @@ const Dashboard = () => {
         {currentRepos.length > 0 ? (
           <>
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-              {currentRepos.map((repo, index) => (
+              {currentRepos.map((repo: Repository, index: number) => (
                 <RepoCard key={`${repo.metadata.name}-${index}`} repo={repo} />
               ))}
             </div>
