+++ src/components/Pagination.tsx
@@ -1,0 +1,47 @@
+import { ChevronLeft, ChevronRight } from 'lucide-react';
+
+interface PaginationProps {
+  currentPage: number;
+  totalPages: number;
+  onPageChange: (page: number) => void;
+}
+
+export const Pagination = ({ currentPage, totalPages, onPageChange }: PaginationProps) => {
+  return (
+    <div className="flex items-center justify-center gap-2 mt-8">
+      <button
+        onClick={() => onPageChange(currentPage - 1)}
+        disabled={currentPage === 1}
+        className="p-2 rounded-lg hover:bg-gray-200 disabled:opacity-50 disabled:hover:bg-transparent"
+      >
+        <ChevronLeft size={20} />
+      </button>
+
+      <div className="flex items-center gap-1">
+        {Array.from({ length: totalPages }, (_, i) => i + 1).map((page) => (
+          <button
+            key={page}
+            onClick={() => onPageChange(page)}
+            className={`px-3 py-1 rounded-lg ${
+              currentPage === page ? 'bg-blue-600 text-white' : 'hover:bg-gray-200'
+            }`}
+          >
+            {page}
+          </button>
+        ))}
+      </div>
+
+      <button
+        onClick={() => onPageChange(currentPage + 1)}
+        disabled={currentPage === totalPages}
+        className="p-2 rounded-lg hover:bg-gray-200 disabled:opacity-50 disabled:hover:bg-transparent"
+      >
+        <ChevronRight size={20} />
+      </button>
+
+      <div className="ml-4 text-sm text-gray-600">
+        Page {currentPage} of {totalPages}
+      </div>
+    </div>
+  );
+};
