+++ src/components/SearchAndFilter.tsx
@@ -1,0 +1,56 @@
+import { Search, Filter } from 'lucide-react';
+
+interface SearchAndFilterProps {
+  searchTerm: string;
+  onSearchChange: (term: string) => void;
+  selectedCategories: string[];
+  onCategoryChange: (category: string) => void;
+  categories: string[];
+}
+
+export const SearchAndFilter = ({
+  searchTerm,
+  onSearchChange,
+  selectedCategories,
+  onCategoryChange,
+  categories,
+}: SearchAndFilterProps) => {
+  return (
+    <div className="mb-8 space-y-4">
+      <div className="relative">
+        <Search
+          className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400"
+          size={20}
+        />
+        <input
+          type="text"
+          placeholder="Search projects..."
+          value={searchTerm}
+          onChange={(e) => onSearchChange(e.target.value)}
+          className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+        />
+      </div>
+      <div>
+        <div className="flex items-center gap-2 mb-2">
+          <Filter size={20} className="text-gray-600" />
+          <span className="font-medium text-gray-700">Categories</span>
+        </div>
+        <div className="flex flex-wrap gap-2">
+          {categories.map((category) => (
+            <button
+              key={category}
+              onClick={() => onCategoryChange(category)}
+              className={`px-3 py-1 rounded-full text-sm ${
+                selectedCategories.includes(category)
+                  ? 'bg-blue-600 text-white'
+                  : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
+              }`}
+            >
+              {category}
+            </button>
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+};
